- 좋아요 삭제 시 좋아요 조회를 좋아요 id가 아닌 likesId와 UserId로 하던 로직 좋아요 id로 조회 후 equals로 확인하는 것으로 수정  -> 이유 : index로 검색하는 검색 시 더 빠름

+++ src/main/java/com/gnakkeoyhgnus/noteforios/service/LikesService.java
@@ -8,6 +8,7 @@ import com.gnakkeoyhgnus.noteforios.domain.repository.LikesRepository;
 import com.gnakkeoyhgnus.noteforios.domain.repository.PageShareBoardRepository;
 import com.gnakkeoyhgnus.noteforios.exception.CustomException;
 import com.gnakkeoyhgnus.noteforios.exception.ErrorCode;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -46,8 +47,12 @@ public class LikesService {
 
   @Transactional
   public void deleteLikes(User user, Long likesId) {
-    Likes likes = likesRepository.findByIdAndUserId(likesId, user.getId())
+    Likes likes = likesRepository.findById(likesId)
         .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_LIKES));
+
+    if (!Objects.equals(user.getId(), likes.getUser().getId())) {
+      throw new CustomException(ErrorCode.PERMISSION_DENIED_TO_DELETE);
+    }
 
     likesRepository.delete(likes);
   }
